// src/components/fileDropzone/interface.ts
export interface FileDTO {
  // id?: Number;
  docId: string;
  // alfrescoId: string;
  // name: string;
  // fileSize: Number;
  // mimeType: string;
  // screenName?: string;
  // isDeleted?: boolean;
  // createDate?: Date | null;
  // userId?: string;
  // dipFileId: Number;
  // fileEntryId?: Number;
  // url?: string;

  _id?: string;
  contentId?: string;
  contentType: string;
  orgCode: string;
  name: string;
  extension: string;
  alfrescoId: string;

}